requirements.txt
pyfrc
pyntcore
robotpy-installer
robotpy-ctre
robotpy-hal
robotpy-halsim-gui
robotpy-rev
robotpy-rev-color
robotpy-wpiutil
wpilib
robotpy-wpilib-utilities
pynetconsole
pynetworktables
